Fixed store initialization
`App.store` should be an instance (lowercase `s`), while `App.Store`
should be a class. It was mixed up, resulting in a non-functioning
store.




Also I found the namespace somewhat confusing at first, but it seems
it is named like that in the super `RESTAdapter`. So I added a note
about it.